Moved API calls up to CourseMainPageContainer.vue
